wip migrating saved data to custom classes - vendor models
