--- Example/nrf-mesh/app/src/main/java/no/nordicsemi/android/nrfmeshprovisioner/utils/InterfaceAdapter.java
@@ -1,75 +1,0 @@
-/*
- * Copyright (c) 2018, Nordic Semiconductor
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the
- * documentation and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote products derived from this
- * software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
- * USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package no.nordicsemi.android.nrfmeshprovisioner.utils;
-
-import com.google.gson.JsonDeserializationContext;
-import com.google.gson.JsonDeserializer;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParseException;
-import com.google.gson.JsonSerializationContext;
-import com.google.gson.JsonSerializer;
-
-import java.lang.reflect.Type;
-
-public class InterfaceAdapter<T> implements JsonSerializer<T>, JsonDeserializer<T> {
-    @Override
-    public final JsonElement serialize(final T object, final Type interfaceType, final JsonSerializationContext context) {
-        final JsonObject member = new JsonObject();
-
-        member.addProperty("type", object.getClass().getName());
-
-        member.add("data", context.serialize(object));
-
-        return member;
-    }
-
-    @Override
-    public final T deserialize(final JsonElement elem, final Type interfaceType, final JsonDeserializationContext context)
-            throws JsonParseException {
-        final JsonObject member = (JsonObject) elem;
-        final JsonElement typeString = get(member, "type");
-        final JsonElement data = get(member, "data");
-        final Type actualType = typeForName(typeString);
-
-        return context.deserialize(data, actualType);
-    }
-
-    private Type typeForName(final JsonElement typeElem) {
-        try {
-            return Class.forName(typeElem.getAsString());
-        } catch (ClassNotFoundException e) {
-            throw new JsonParseException(e);
-        }
-    }
-
-    private JsonElement get(final JsonObject wrapper, final String memberName) {
-        final JsonElement elem = wrapper.get(memberName);
-
-        if (elem == null) {
-            throw new JsonParseException(
-                    "no '" + memberName + "' member found in json file.");
-        }
-        return elem;
-    }
-}
